feat: added singleGuild block

+++ src/bot/listeners/channels/ChannelDelete.ts
@@ -8,6 +8,7 @@ export class ChannelDeleteListener extends Listener {
 	public async run(channel: DMChannel | GuildChannel) {
 		const { client } = this.container;
 
+		if (channel.type !== "DM" && channel.guildId !== client.constants.guild) return;
 		if (!channel.isText()) return;
 		const ticket = await client.prisma.ticket.findFirst({ where: { channel: channel.id } });
 		if (ticket) return await client.prisma.ticket.delete({ where: { caseId: ticket.caseId } });
